Cleanup stuff

+++ ModLoaderClient/Api/GameLobbyWrapper.cs
@@ -1,48 +1,55 @@
 using AmongUs.Api;
 
-namespace AmongUs.Client.Loader.Api 
+namespace AmongUs.Client.Loader.Api
 {
-	public readonly struct GameLobbyWrapper : IGameLobby 
+	public readonly struct GameLobbyWrapper : IGameLobby
 	{
 		private GameStartManager Original { get; }
 
-		public float CountDownTimer 
+		public float CountDownTimer
 		{
 			get => Original.AKLOKGOIKHP;
 			set => Original.AKLOKGOIKHP = value;
 		}
 
-		public int MinPlayers {
+		public int MinPlayers
+		{
 			get => Original.MinPlayers;
 			set => Original.MinPlayers = value;
 		}
-		
-		public int LastPlayerCount { 
+
+		public int LastPlayerCount
+		{
 			get => Original.OHFFOOPPAFK;
 			set => Original.OHFFOOPPAFK = value;
 		}
 
-		public StartingState StartState {
-			get => (StartingState)(int)Original.FEFLPKBHACE;
-			set => Original.FEFLPKBHACE = (GameStartManager.PNMJGJHIGIN)value;
+		public StartingState StartState
+		{
+			get => (StartingState) Original.FEFLPKBHACE;
+			set => Original.FEFLPKBHACE = (GameStartManager.PNMJGJHIGIN) value;
 		}
 
-		public string GameStartText {
+		public string GameStartText
+		{
 			get => Original.GameStartText.Text;
 			set => Original.GameStartText.Text = value;
 		}
 
-		public string GameRoomName {
+		public string GameRoomName
+		{
 			get => Original.GameRoomName.Text;
 			set => Original.GameRoomName.Text = value;
 		}
 
-		public string PlayerCounter {
+		public string PlayerCounter
+		{
 			get => Original.PlayerCounter.Text;
 			set => Original.PlayerCounter.Text = value;
 		}
 
-		public void ResetStartState() {
+		public void ResetStartState()
+		{
 			Original.ResetStartState();
 		}
